Separate integrations code to different files
