+++ src/gurubase-backend/backend/core/integrations/github_strategy.py
@@ -1,0 +1,105 @@
+
+import logging
+from core.integrations.helpers import IntegrationError
+from core.integrations.strategy import IntegrationStrategy
+from core.models import GuruType, Integration
+
+logger = logging.getLogger(__name__)
+
+class GitHubStrategy(IntegrationStrategy):
+    def __init__(self, integration: 'Integration' = None):
+        from core.github.app_handler import GithubAppHandler
+        super().__init__(integration)
+        self.github_handler = GithubAppHandler(integration)
+
+    def _fetch_repositories(self, installation_id: str, client_id: str = None, private_key: str = None) -> list:
+        """Fetch repositories for a GitHub installation"""
+        try:
+            return self.github_handler.fetch_repositories(installation_id, client_id, private_key)
+        except Exception as e:
+            logger.error(f"Error fetching GitHub repositories: {e}", exc_info=True)
+            return []
+        
+    def _fetch_installation(self, installation_id: str, client_id: str = None, private_key: str = None) -> dict:
+        """Fetch installation details for a GitHub installation"""
+        try:
+            return self.github_handler.get_installation(installation_id, client_id, private_key)
+        except Exception as e:
+            logger.error(f"Error fetching GitHub installation: {e}", exc_info=True)
+            return {}
+
+    def exchange_token(self, code: str) -> dict:
+        """For GitHub, we don't exchange a code. Instead, we use the installation_id as the external_id."""
+        raise NotImplementedError("GitHub integration does not use code exchange")
+
+    def get_external_id(self, token_response: dict) -> str:
+        """For GitHub, the external_id is the installation_id"""
+        return token_response.get('installation_id')
+
+    def get_workspace_name(self, installation_id: str) -> str:
+        """For GitHub, we use the repository names as the workspace name"""
+        installation = self._fetch_installation(installation_id)
+        if not installation:
+            return f"GitHub Installation {installation_id}"
+            
+        return installation.get('account', {}).get('login')
+
+    def list_channels(self, installation_id: str = None, client_id: str = None, private_key: str = None) -> list:
+        """For GitHub, we return repositories as channels"""
+        repo_names = self._fetch_repositories(installation_id or self.get_integration().external_id, client_id, private_key)
+        return [{'id': name, 'name': name, 'mode': 'auto'} for name in repo_names]
+
+    def send_test_message(self, channel_id: str) -> bool:
+        """GitHub doesn't support test messages"""
+        return True
+
+    def revoke_access_token(self) -> None:
+        """Delete the installation"""
+        integration = self.get_integration()
+        self.github_handler.delete_installation(integration.external_id)
+
+    def refresh_access_token(self, refresh_token: str) -> dict:
+        """GitHub doesn't support token refresh"""
+        raise NotImplementedError("GitHub tokens don't expire and can't be refreshed")
+
+    def fetch_workspace_details(self, installation_id: str, client_id: str = None, private_key: str = None) -> dict:
+        """For GitHub, we use the installation_id as both external_id and workspace name"""
+        installation = self._fetch_installation(installation_id, client_id, private_key)
+        if not installation:
+            self.github_handler.clear_redis_cache()
+            raise ValueError(f"GitHub installation not found for ID: {installation_id}")
+        workspace_name = installation.get('account', {}).get('login')
+        if not workspace_name:
+            workspace_name = f"GitHub Installation {installation_id}"
+        
+        return {
+            'external_id': installation_id,  # bot_token is actually installation_id in this case
+            'workspace_name': workspace_name
+        }
+
+    def get_type(self) -> str:
+        return 'GITHUB'
+
+    def create_integration(self, installation_id: str, guru_type: GuruType) -> Integration:
+        """Create GitHub integration with the installation ID"""
+        # Check if integration already exists for this type and external_id
+        if Integration.objects.filter(type=self.get_type(), external_id=installation_id).exists():
+            logger.error(f"Integration for {self.get_type()} with ID {installation_id} already exists")
+            raise IntegrationError(f"This integration type is already connected to this guru. Please disconnect the existing integration before connecting a new one.")
+        
+        try:
+            # Fetch repository names for workspace name
+            workspace_name = self.get_workspace_name(installation_id)
+            channels = self.list_channels(installation_id)
+            
+            return Integration.objects.create(
+                type=self.get_type(),
+                external_id=installation_id,
+                guru_type=guru_type,
+                access_token=installation_id,  # For GitHub, we use installation_id as the access_token
+                workspace_name=workspace_name,
+                channels=channels
+            )
+        except Exception as e:
+            logger.error(f"Error creating GitHub integration: {e}", exc_info=True)
+            raise IntegrationError(f"Error creating GitHub integration. Please try again. If the problem persists, please contact support.")
