+++ src/gurubase-backend/backend/core/integrations/rest_commands.py
@@ -4,8 +4,9 @@ from rest_framework.response import Response
 from rest_framework import status
 from django.conf import settings
 import logging
-from .models import Integration, GuruType
-from .integrations import IntegrationFactory, IntegrationError
+from core.models import Integration, GuruType
+from .helpers import IntegrationError
+from .factory import IntegrationFactory
 
 logger = logging.getLogger(__name__)
